requirements.txt
click==6.7
dnspython==1.15.0
Flask==0.12.2
Flask-Cors==3.0.3
Flask-HTTPAuth==3.2.3
Flask-Redis==0.3.0
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
kafka-python==1.4.3
MarkupSafe==1.0
passlib==1.7.1
redis==2.10.6
six==1.11.0
SQLAlchemy==1.2.7
Werkzeug==0.14.1
